+++ docs/changes.md
@@ -7,6 +7,7 @@
 - Added 'FrameId' to time manager.
 - Added 'SyncUpdateWith' to RigidBody to fix rigid body updating + camera jerking artifact.
 - Added 'AutoUpdate' property to camera, for better control over when it updates.
+- Added basic renderer with material override, so that animators could be used on sprites as well.
 
 ## 0.1.0.7